nlp_unit/requirements.txt
alabaster==0.7.10
appnope==0.1.0
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
Babel==2.5.1
boto3==1.4.7
botocore==1.7.18
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
constantly==15.1.0
cssselect==1.0.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
docutils==0.14
first==2.0.1
ftfy==4.4.3
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
howdoi==1.1.9
html5lib==0.999999999
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipython==6.2.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
klein==17.2.0
lazy-object-proxy==1.3.1
lxml==4.0.0
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mitie==0.2.0
msgpack-python==0.4.8
murmurhash==0.26.4
neovim==0.1.13
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.5
pandas==0.21.1
parso==0.1.0
pathlib==1.0.1
pew==0.1.26
pexpect==4.2.1
pickleshare==0.7.4
pip-tools==1.9.0
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pylint==1.7.4
pymongo==3.5.1
pyparsing==2.2.0
pyquery==1.2.17
pytest==3.2.3
python-crfsuite==0.9.5
python-dateutil==2.6.1
python-twitter==3.3
pythonz-bd==1.11.4
pytz==2017.2
rasa-nlu==0.10.0
regex==2017.9.23
requests==2.18.4
requests-cache==0.4.13
requests-oauthlib==0.8.0
resumable-urlretrieve==0.1.5
s3transfer==0.1.11
scikit-learn==0.19.0
scipy==0.19.1
semver==2.7.8
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
sklearn-crfsuite==0.3.6
snowballstemmer==1.2.1
spacy==1.9.0
Sphinx==1.6.4
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
stop-words==2015.2.23.1
tabulate==0.8.1
termcolor==1.1.0
thinc==6.5.2
toolz==0.8.2
tqdm==4.17.1
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
ujson==1.35
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.11
yapf==0.19.0
zope.interface==4.4.3
